refactor: move window.hpp to desktop/
