+++ src/render/decorations/IHyprWindowDecoration.cpp
@@ -1,6 +1,6 @@
 #include "IHyprWindowDecoration.hpp"
 
-#include "../../Window.hpp"
+class CWindow;
 
 IHyprWindowDecoration::IHyprWindowDecoration(CWindow* pWindow) {
     m_pWindow = pWindow;
